refactor(recorder): Move recorder related functions to recorder.py
